backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
websockets
google-genai
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
bcrypt
pytest
httpx
ruff
sqlmodel
alembic
